Begin deprecation of goofiboot (T4235)
Switch clr-boot-manager to now use systemd-boot by default instead of
goofiboot. Existing installs will be unaffected by the transition in
a negative capacity, save for the wasted 84k on the ESP (nothing we
can truly do about this.)

EFI Vars for "Linux Boot Manager" will continue booting the old stub
and the default entry (which is what the majority are using) will boot
the new stub. Users wishing to manually transition can use bootctl to
reinstall the stub.

When Solus transitions to clr-boot-manager 2.0 (not scheduled to happen
immediately until we've done some work upstream to support efivar
management in a more generic fashion) we'll automatically handle the
var switch.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>